refactor: RoundRange enum 클래스 생성을 통한 Round 클래스 리팩토링
* style: domain 패키지 구조 세분화 진행
